Arrumando bug funcionario e criando o atualizar do setor

+++ WebContent/resources/js/modal-setor.js
@@ -3,7 +3,9 @@
 		data:function(){
 			return {
 				listaSetores: [],
-				nomeSetor: null
+				nomeSetor: null,
+				id: null,
+				mensagem: null
 			}
 		},
 		template:`
@@ -15,6 +17,12 @@
 								<button type="button" @click="fecharModal($event)" class="close"><span >&times;</span></button>
 							</div>
 							<div class="modal-body">
+								<div v-if="mensagem" class="alert alert-primary alert-dismissible fade show" role="alert">
+									<span>{{mensagem}}</span>
+									<button type="button" @click="fecharMensagemSetor()" class="close">
+										<span>&times;</span>
+									</button>
+								</div>
 								<div class="col">
 									<h3>Cadastrar um setor</h3>
 									<div class="form-group input-setor">
@@ -34,7 +42,7 @@
 											<td>{{setor.id}}</td>
 											<td>{{setor.nome}}</td>
 											<td>
-												<span @click="editarSetor(setor.nome)">&#9998;</span>
+												<span @click="editarSetor(setor.id,setor.nome)">&#9998;</span>
 												<span @click="excluirSetor(setor.id)">&#10006;</span>
 											</td>
 										</tr>
@@ -45,7 +53,6 @@
 							</div>
 							<div class="modal-footer">
 								<button type="button" class="btn btn-default" @click="fecharModal($event)">Fechar</button>
-								<button type="button" class="btn btn-primary" @click="excluir()">Confirmar</button>
 							</div>
 					</div>
 				</div>
@@ -56,13 +63,22 @@
 			 },
 			methods:{
 				excluirSetor(id){
-					axios.delete("/funcionarios/rs/setor/"+id)
-					.then(response => {
-						this.buscaSetores();
-					}).catch(function (error) {
-						vm.mostraAlertaErro("Erro interno", "Não foi listar natureza de serviços");
-					}).finally(function() {
-					});
+					axios.get("/funcionarios/rs/funcionarios/setor/"+id).then(funcionarios =>{
+						console.log(funcionarios);
+						if(funcionarios.data.length){
+							this.mensagem = "Esse setor não pode ser excluir, pois tem funcionario vinculado";
+						}else{
+							axios.delete("/funcionarios/rs/setor/"+id)
+							.then(response => {
+								this.buscaSetores();
+								this.mensagem = "Setor excluido com sucesso";
+							}).catch(function (error) {
+								this.mostraAlertaErro("Erro interno", "Não foi deletar o setor");
+							}).finally(function() {
+							});
+							
+						}
+					})
 				},
 				buscaSetores(){
 					axios.get("/funcionarios/rs/setor")
@@ -74,15 +90,33 @@
 				},
 				salvarSetor(){
 					let setor = {
+							id:this.id,
 							nome: this.nomeSetor
 					}
-					axios.post("/funcionarios/rs/setor", setor).then(response =>{
-						this.limparCampos();
-						this.buscaSetores();
-						this.mensagem = "Funcionario salvo com sucesso";
-					}).catch(function (error){
-						this.mostraAlertaErro("Erro interno");
-					});
+					if(this.id != null){
+						axios.put("/funcionarios/rs/setor/"+this.id, setor).then(response =>{
+//							this.limparCampos();
+							this.id = null;
+							this.nomeSetor = null;
+							this.mensagem = "setor atualizado com sucesso";
+							this.buscaSetores();
+						}).catch(function (error){
+							this.mostraAlertaErro(error);
+						});
+					
+					}else{
+						if(!this.nomeSetor){
+							this.mensagem = "O nome do setor não pode ficar vazio";
+						}else{
+							axios.post("/funcionarios/rs/setor", setor).then(response =>{
+								this.limparCampos();
+								this.buscaSetores();
+								this.mensagem = "Setor salvo com sucesso";
+							}).catch(function (error){
+								this.mostraAlertaErro("Erro interno");
+							});
+						}
+					}
 				},
 				limparCampos(){
 					this.nomeSetor = null;
@@ -92,7 +126,18 @@
 				},
 				fecharModal(evento){
 					this.$parent.fecharModal();
+				},
+				editarSetor(id,nome){
+					console.log(nome);
+					this.id = id;
+					this.nomeSetor = nome; 
+				},
+				
+				fecharMensagemSetor(){
+					this.mensagem = null;
 				}
+				
+				
 				
 			}
 	});
